issue-1046 Renaming module (and some indentation fixes)
