+++ src/app/modules/logging-management/logging-management.module.ts
@@ -5,24 +5,24 @@ import {LayoutModule} from '../../shared/layout/layout.module';
 import {MaterialModule} from '../../shared/material.module';
 import {FormHelperModule} from '../../shared/form/form-helper.module';
 import {LoggingService} from './services/logging.service';
-import {LoggingRoutingModule} from './logging-routing.module';
+import {LoggingManagementRouting} from './logging-management.routing';
 import {UpdateLoggerDialogComponent} from './dialogs/update-logger/update-logger.dialog.component';
 
 @NgModule({
-  declarations: [
-    LoggingComponent,
-    UpdateLoggerDialogComponent
-  ],
-  imports: [
-    CommonModule,
-    LayoutModule,
-    MaterialModule,
-    FormHelperModule,
-    LoggingRoutingModule
-  ],
-  providers: [
-    LoggingService
-  ]
+    declarations: [
+        LoggingComponent,
+        UpdateLoggerDialogComponent
+    ],
+    imports: [
+        CommonModule,
+        LayoutModule,
+        MaterialModule,
+        FormHelperModule,
+        LoggingManagementRouting
+    ],
+    providers: [
+        LoggingService
+    ]
 })
-export class LoggingModule {
+export class LoggingManagementModule {
 }
