+++ src/app/modules/logging-management/pages/logging/logging.component.scss
