+++ src/app/modules/logging-management/services/logging.service.spec.ts
@@ -1,0 +1,12 @@
+import {TestBed} from '@angular/core/testing';
+
+import {LoggingService} from './logging.service';
+
+describe('LoggingService', () => {
+    beforeEach(() => TestBed.configureTestingModule({}));
+
+    it('should be created', () => {
+        const service: LoggingService = TestBed.get(LoggingService);
+        expect(service).toBeTruthy();
+    });
+});
