--- src/app/modules/logging/services/logging.service.spec.ts
@@ -1,12 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { LoggingService } from './logging.service';
-
-describe('LoggingService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
-
-  it('should be created', () => {
-    const service: LoggingService = TestBed.get(LoggingService);
-    expect(service).toBeTruthy();
-  });
-});
